Remove StreamForwarder Tests from E2E
