+++ scripts/test/runtests.ps1
@@ -9,6 +9,7 @@ $TestBinRoot = "$RepoRoot\artifacts\tests"
 
 $TestProjects = @(
     "E2E",
+    "StreamForwarderTests"
     "Microsoft.DotNet.Tools.Publish.Tests"
 )
